tracks_sold_usa
The first query returns each genre with the number of tracks sold in the USA in absolute numbers and in percentages. The top 3 selling genres in the USA are rock, alternative & punk, and metal. 

The second query returns the top 10 selling artists with the top 3 selling genres in the USA - rock, alternative & punk, and metal. The top 5 artists are Iron Maiden, Led Zeppelin, Metallica, U2, and Deep Purple, all of which are rock genres except for Metallica, which is a metal genre.